Cleaning up paral_help in schema
-The \' were not very useful, setting on using *var suffix for naming convention, although this is not really used, only for documentation purposes.

-Note that this param_help serves a dual function!! Documentation/hint for user
in using the get/set methods. Gets kind of tricky with these nested dicts.The second purpose is for mapping the command line switches to the dictionary. Some of the dicts are dynamic so thee only way to feed them from the cmdline is to insert the keys int eh arg value. The param_help assists in figuring out how these keys should be mapped to the dictionary structure.