Document OpenROAD magic numbers in sc_apr.tcl
I have a couple otherwise independent commits I'm planning to make that
need this structure laid out, so submitting this in advance to avoid
merge issues. This commit provides us the ability to adjust OpenROAD
constants based on platform, and documents what's going on/adds a
warning to try and avoid related confusion.