floorplan: update API for consistency
All place_* functions now take in lists of instances, a starting
position, pitch, and direction. `place_pins` is slightly unique in that
it takes in a side to place the pins along.

In addition, instead of having a units argument, everything is
automatically snapped to a grid specified in the constructor. The user
may scale dimensions using values captured in the Floorplan object. When
necessary, the user is responsible for snapping values to a coarser grid
using helper functions in the library.