Don't use deref(p) as RHS of any assignment
Although we can get away with it in some cases, this is super dangerous
and to be extra careful we just shouldn't do it. Even though the
assignment performs a copy by value, the problem is some of the LEF
parser objects have members that are pointers to other objects, and they
free these pointers when their destructor is called. Therefore, these
internal objects will be double free()-d when the functionr eturns.