Modify insert_vias for non-orthogonal intersection
Previously, insert_vias treated each wire as a 1D line, found
intersections as points where those lines intersected, and then used the
widths of the wires to determine the intersection area.

The problem with this approach is it only works for intersections where
the top and bottom wires are orthogonal. However, I realized this isn't
always the case - in particular, the wires that connect ZeroSoC's power
ring to the power pads do not fit this assumption.

I changed the approach to instead treat each wire as a rectangle, and
fill the regions where these rectangles intersect. This is more similar
to how OpenROAD's PDN generation works, is more general, and has the
positive side-effect of actually reducing the geometric complexity.

OpenROAD's algorithm for generating via arrays takes into consideration
the direction of the top and bottom layer. Previously we determined
these directions based on the sizes of the wires, but now we just use
the preferred routing direction of the layers, which is also how
OpenROAD does it.