Fix user-provided modules via SCPATH
Since we reorganized the directory structure to put our build in tools,
flows, etc. under siliconcompiler/siliconcompiler, we had to prefix the
dynamic import searchdirs with `siliconcompiler.`. However, this seemed
to break the ability for user's to provide their own dynamic modules for
tools, etc.

By adding `siliconcompiler/siliconcompiler/` to the sys.path, and
removing `siliconcompiler.` from the loadtool/pdk/flow routines, we go
back to the old behavior where we'll search both the SC builtins and
user provided directories.