Refactor logging to fix multiprocessing bug
It turns out that the root cause of the multiprocessing bug is that at
least one of the objects involved in logging is not serializable, and
therefore the chip object itself wasn't serializable. However, it needs
to be in order for us to call _runstep in a new process, since the chip
object itself needs to be communicated to that process somehow.

In order to work around this, I created two helper functions,
_init_logger() and _deinit_logger() that will initialize and
deinitialize these objects. The code now deinitializes the logger before
spawning processes, and each thread will then initialize its own logger
as its very first step. We then reinitialize the logger of the main
thread when control returns.

As part of this PR, I also slightly refactored how the logger was
getting initialized and added comments to clarify some things that
confused me at first.