Fix test_gcd_cli()
Now that we search for built-in files relative to
siliconcompiler/siliconcompiler/, we can't find the sources files
specified in run.sh anywhere. To work around this, switch to just
building the script from a string template using absolute paths. I think
the CLI API has been stable enough that this shouldn't be a big deal.

It does mean that we won't get automated testing of run.sh, but we can
probably get rid of that in favor of gcd.py.