Add test fixture for instantiating GCD chip
This can save us a lot of boilerplate in our test definitions, and cut
down on the refactoring involved in API changes.

In order to support running a test directly as a main module, I create
two mirror copies the gcd_chip() function -- one as a regular Python
function in tests/utils.py, and the other as a proper pytest fixture
that calls this function. We can then choose whether to import
tests.utils or tests.fixtures based on whether the test is running as
main or not.