Replace shutil.copytree() with wrapper function
We use the "dirs_exist_ok" argument to copytree() a lot, but this isn't
supported prior to Python 3.8. This commit hardcodes a more modern
version of shutil.copytree() into a utils file, and adds a wrapper that
will call this hardcoded version or the system version based on the
Python version.