Fix client.py for authenticated/encrypted jobs
Sorry if my use of 'job_nameid' as a variable name has become confusing; it came from when we separated job names and IDs in the schema, and I might need to clean up the `client.py` file a bit. This bug probably didn't show up in our CI tests because we run the authenticated workflow during the daily tests, not the quick tests.

Maybe, considering our plans to launch with beta accounts, we should run the unauthenticated remote flow as a daily test, and the authenticated one as a quick test.