Deduplicate names of imported files
Since we copy all imported files into a single directory, we previously
did not support importing more than one file with the same base
filename. However, we are starting to run into use-cases where this
might occur.

This change allows for importing files with duplicate basenames by
appending a hash of the file's full path to the file's basename when
copying it into the import/ directory. find_files() is then responsible
for reconstructing this filename to properly resolve its location in
import/.