Automate most submodule management
This automates clones of design/tool submodules by coupling them to the
places where the submodules are actually needed. This prevents us from
having to do a fine-grained selection of which submodules are needed in
CI scripts, while still avoiding a complete download of our many
submodules.

The most intricate part are the new test fixtures: these will clone
design submodules as needed by tests and return a path to the submodule.
If the submodule is already cloned there will be no effect, and the
fixture will not attempt to clone the submodule more than once per test
session, since the 'session' scope is set.