Skip test_env on macOS/Windows
These platforms use the multiprocessing "spawn" start method by default,
which doesn't propagate the mocked _runtask(), breaking this test. I
manually verified that the environment gets set in _runtask() in macOS.
Going forward, it might be a good idea to find another way to write this
test to ensure we don't have regressions on these platforms.