Fix check_manifest() bug in remote_preprocess()
Another bug hidden by the check_manifest() steplist issue. In this case,
only the local step is set up before remote_preprocess() calls
_runtask(), so we need to override the steplist to force it to only
check the local step. This feels a bit hacky, but it shouldn't be
dangerous since we already overwrite the steplist for the remote run
immediately after.

I also snuck in a separate bugfix - it seems like the _runtask() call
needs to be outside of the if statement, otherwise the local step will
not run if the local step runs a builtin function.