Fix bug causing [arg, index] to not be respected
Since we clobber [arg, index] when running the tool setup loop, using it
later to determine whether or not to only run a single index for a step
doesn't work. Precomputing a per-step indexlist avoids this problem and
also makes the code cleaner.

This also fixes an additional uncaught bug where the indexlist parameter
wouldn't be respected by run().