Fix up manifest merging logic at end of run()
This code was previously based on an assumption that the index of the
final task to run is 0. We've decided that all flowgraphs should have a
single merged leaf, but this assumption is still problematic. For
example, this logic breaks in the following circumstances:

- A user specifies a subset of a flowgraph to run that has multiple
  parallel steps, but does not include the merge step
- A user runs a single task with [arg, step]/[arg, index] where the
  index is not 0. This one in particular comes up when we're running a
  flowgraph using Slurm.

The solution is to use the indexlist to ensure that we check manifests
for all indices associated with our final step.