Fix Read The Docs build
Since we added read_lef() to core, core.py imports leflib. This breaks
our Read The Docs build, since it can't handle native extensions (but it
needs to run SC to do auto-doc generation). This is a bit of a hack,
but this change lets you run SC without building the LEF parser (and
gives you an ugly warning).