Replace direct cfg merge w/ partial read_manifest
There are a few problems with doing this direct update, namely that it
loses previous task info and that it wipes out "default" keys, making it
possibly to get an error when trying to update these dictionaries in the
future.

Creating a "partial" read_manifest() option that only reads the keypaths
we need gets us the majority of the performance benefit without these
performance pitfalls.