Only run dynamic manifest checks during _runtask()
This fixes a performance problem where each task has to incur the
overhead of checking an entire flowgraph for really large graphs. Since
those sorts of things are static, they only need to be checked once in
run(), and the check_manifest() calls inside _runtask() can focus on
dynamic checks.