Fix bug introduced by #1145
This commit takes a lighter approach to fixing the immediate bug that we
attempted to fix with #1145, which is to ensure that tool setup persists
across re-runs using steplist. Reading in the entire manifest before a
run overwrites boolean options (such as resume), which can lead to
unintuitive behavior.

I still think there are better fixes down the road, but this seems like
a better fix to go with in the meantime.