schema_cfg: store native types in dictionary
There's no reason to store everything as a string - it seems simpler to
keep everything as their native Python type.

The only type that changes when converting to/from JSON are tuples,
which get converted to lists. We can keep them as lists in the schema
for the sake of normalization.