Add check for unresolved filepaths to _abspath()
When dumping a manifest, some paths will not be able to be resolved, and
find_files() will return lists with "None" entries. This seemed to work
okay with the old set() method, but violates our new typechecks. In
these cases the behavior was for the paths to not get included in the
manifest, so it seems okay to write an empty list instead.