core: refactor report to reduce direct manifest access
We can pass data directly in rather than having to hardcode dictionary
access in the Jinja template.

To facilitate code reuse, this commit also refactors to summary() to
separate out its data extraction and display logic.

Unfortunately it's not quite possible to completely decouple the
manifest structure from the report, since the report lets you modify and
save copies of the manifest.