[Tools][opensta] Updated SDF Interface Based on Peter Feedback
Made it so SPEF and SDF are mutually exclusive when being provided to
OpenSTA.

Made it so the SDF had to be passed into the input directory.

Added a testing tool which will copy any file into the output directory
so the SDF file can be provided to OpenSTA without being generated.