[Tools][VPR] Added Option to Generate Post Implementation Netlist
Similar to the "max_router_iterations" option in the VTR flow, created
an option that can be passed into the VPR tool to have the route stage
generate a post-implementation netlist. This option will put the
generated files into the output directory in a form that can be accepted
into the OpenSTA tool.