[Tools][VPR] Post-Implementation Netlist File Generation
The VPR tool driver decides whether to use timing analysis or not based
on if the user provides an SDC file or not. This is a problem for adding
an OpenSTA timing node after routing since a post-implementation netlist
will not be generated if timing is off (timing information is currently
needed to generate this file).

Resolved this issue by having the VPR tool driver always generate the
post-implementation netlist files, but just makes them blank if timing
analysis is off. The OpenSTA tool driver can then skip itself if it
recognizes that blank files were passed into it. This allows the timing
analysis tool to be integrated cleanly into the FPGA flow.