[SQUASH ME] Removed Changes to FPGA Flow
Found that this cannot be a part of the FPGA flow so long as the FPGA
does not have the asic keys defined.

This overall PR is still useful for fixing the issue with the SDC file
not being present, but we must use a target in Logik to use this
feature.