[Schedular] Do Not Forward non-Required Outputs
When a schedular node is skipped, it currently forwards all of its
inputs to the output directory. If the tool was not expecting these
files in the output directory, this can raise an issue.

To fix this issue, when a node is being skipped, we now only forward the
input files which were required to be output files.